HOT-FIX: metadata, package updates and more
- move metadata to be static rather than generating through constructMetadata - turns out this is uneeded as nextjs handles missing fields (where in the past it might have not)
- packages updated to latest
- general code clean up
- fix up robots to allow sign-in and sign-up
