+++ apps/web/package.json
@@ -14,76 +14,76 @@
   "dependencies": {
     "@buzztrip/backend": "workspace:*",
     "@buzztrip/transactional": "workspace:*",
-    "@clerk/nextjs": "^6.22.0",
-    "@convex-dev/react-query": "^0.0.0-alpha.8",
-    "@headless-tree/core": "^1.0.1",
-    "@headless-tree/react": "^1.0.1",
+    "@clerk/nextjs": "^6.23.0",
+    "@convex-dev/react-query": "^0.0.0-alpha.11",
+    "@headless-tree/core": "^1.2.0",
+    "@headless-tree/react": "^1.2.0",
     "@hookform/resolvers": "^5.1.1",
     "@mapbox/search-js-react": "^1.1.0",
     "@mdx-js/loader": "^3.1.0",
     "@mdx-js/react": "^3.1.0",
-    "@next/mdx": "^15.3.3",
-    "@polar-sh/sdk": "^0.33.1",
+    "@next/mdx": "^15.3.4",
+    "@polar-sh/sdk": "^0.33.2",
     "@radix-ui/react-icons": "^1.3.2",
-    "@sentry/nextjs": "^9.29.0",
+    "@sentry/nextjs": "^9.31.0",
     "@t3-oss/env-nextjs": "^0.13.8",
     "@tailwindcss/postcss": "^4.1.10",
-    "@tanstack/react-query": "^5.80.7",
-    "@tanstack/react-query-devtools": "^5.80.7",
+    "@tanstack/react-query": "^5.81.2",
+    "@tanstack/react-query-devtools": "^5.81.2",
     "@uidotdev/usehooks": "^2.4.1",
-    "@uiw/color-convert": "^2.5.5",
-    "@uiw/react-color-colorful": "^2.5.5",
+    "@uiw/color-convert": "^2.6.0",
+    "@uiw/react-color-colorful": "^2.6.0",
     "@vis.gl/react-google-maps": "^1.5.3",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
-    "convex": "^1.24.8",
+    "convex": "^1.25.0",
     "date-fns": "^4.1.0",
     "embla-carousel-react": "^8.6.0",
     "input-otp": "^1.4.2",
-    "lucide-react": "^0.515.0",
-    "mapbox-gl": "^3.12.0",
-    "motion": "^12.18.1",
-    "next": "^15.3.3",
-    "next-safe-action": "^8.0.2",
+    "lucide-react": "^0.523.0",
+    "mapbox-gl": "^3.13.0",
+    "motion": "^12.19.1",
+    "next": "^15.3.4",
+    "next-safe-action": "^8.0.4",
     "next-themes": "^0.4.6",
     "nuqs": "^2.4.3",
-    "posthog-js": "^1.252.0",
+    "posthog-js": "^1.255.1",
     "radix-ui": "^1.4.2",
     "react": "^19.1.0",
     "react-day-picker": "9.7.0",
     "react-dom": "^19.1.0",
-    "react-hook-form": "^7.57.0",
+    "react-hook-form": "^7.58.1",
     "react-hotkeys-hook": "^5.1.0",
     "react-map-gl": "^8.0.4",
-    "react-resizable-panels": "^3.0.2",
+    "react-resizable-panels": "^3.0.3",
     "react-scan": "^0.3.4",
-    "recharts": "^2.15.3",
+    "recharts": "^3.0.0",
     "remark-gfm": "^4.0.1",
     "sonner": "^2.0.5",
     "svix": "^1.67.0",
     "tailwind-merge": "^3.3.1",
     "tailwindcss-animate": "^1.0.7",
     "vaul": "^1.1.2",
-    "zod": "^3.25.64",
+    "zod": "^3.25.67",
     "zustand": "^5.0.5"
   },
   "devDependencies": {
     "@buzztrip/tsconfig": "workspace:*",
     "@eslint/eslintrc": "^3.3.1",
-    "@tanstack/eslint-plugin-query": "^5.78.0",
+    "@tanstack/eslint-plugin-query": "^5.81.2",
     "@types/eslint-config-prettier": "~6.11.3",
     "@types/google.maps": "^3.58.1",
     "@types/mapbox-gl": "^3.4.1",
     "@types/mdx": "^2.0.13",
-    "@types/node": "^24.0.1",
+    "@types/node": "^24.0.4",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "autoprefixer": "^10.4.21",
     "eslint": "^9.29.0",
-    "eslint-config-next": "^15.3.3",
+    "eslint-config-next": "^15.3.4",
     "eslint-config-prettier": "^10.1.5",
-    "postcss": "^8.5.5",
+    "postcss": "^8.5.6",
     "tailwindcss": "^4.1.10",
     "typescript": "^5.8.3"
   }
