+++ package.json
@@ -14,9 +14,9 @@
     "@buzztrip/tsconfig": "workspace:*",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
-    "prettier": "^3.5.3",
+    "prettier": "^3.6.1",
     "prettier-plugin-organize-imports": "^4.1.0",
-    "prettier-plugin-tailwindcss": "^0.6.12",
+    "prettier-plugin-tailwindcss": "^0.6.13",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "turbo": "^2.5.4",
