+++ packages/backend/convex/maps/index.ts
@@ -1,4 +1,5 @@
 import { zid } from "convex-helpers/server/zod";
+import { uppercaseFirstLetter } from "../../helpers";
 import type { UserMap } from "../../types";
 import {
   mapsEditSchema,
@@ -114,7 +115,7 @@ export const createMap = authedMutation({
 
     const mapId = await ctx.db.insert("maps", {
       ...map,
-      title: `${map.title.charAt(0).toUpperCase()} ${map.title.slice(1)}`,
+      title: uppercaseFirstLetter(map.title),
       owner_id: ctx.user._id,
       mapTypeId: map.mapTypeId ?? "hybrid",
     });
@@ -147,6 +148,7 @@ export const updateMap = authedMutation({
   handler: async (ctx, args) => {
     await ctx.db.patch(args.mapId, {
       ...args.map,
+      title: uppercaseFirstLetter(args.map.title),
       updatedAt: new Date().toISOString(),
     });
   },
